reconfigure: refactor code structure
