+++ lib/workers/repository/reconfigure/utils.ts
@@ -1,0 +1,3 @@
+export function getReconfigureBranchName(prefix: string): string {
+  return `${prefix}reconfigure`;
+}
